requirements.txt
apache-beam==2.34.0
autopep8==1.6.0
avro-python3==1.9.2.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
crcmod==1.7
dill==0.3.1.1
docopt==0.6.2
fastavro==1.4.7
fasteners==0.16.3
future==0.18.2
google-api-core==1.31.4
google-apitools==0.5.31
google-auth==1.35.0
google-cloud-bigquery==2.30.1
google-cloud-bigquery-storage==2.10.0
google-cloud-bigtable==1.7.0
google-cloud-core==1.7.2
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-firestore==2.3.4
google-cloud-language==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-recommendations-ai==0.2.0
google-cloud-spanner==1.19.1
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
grpcio-gcp==0.2.2
hdfs==2.6.0
httplib2==0.19.1
idna==3.3
libcst==0.3.21
mypy-extensions==0.4.3
numpy==1.20.3
oauth2client==4.1.3
orjson==3.6.4
packaging==21.2
proto-plus==1.19.8
protobuf==3.19.1
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydot==1.4.2
pymongo==3.12.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
wincertstore==0.2
